requirements.txt
bs4
fpdf
pillow
cloudscraper
flask
reportlab
flask-cors
